# public/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: GhostClass API
  version: 4.5.0
  description: |
    **GhostClass API** provides endpoints for authentication, profile synchronization,
    attendance integrations with EzyGo, telemetry, and build provenance.

    ## Authentication

    The API supports two authentication methods:

    1. **Bearer Token (CRON_SECRET)**: For automated cron jobs
       - Add `Authorization: Bearer <CRON_SECRET>` header
       - Used by automated sync processes

    2. **Session Cookie (SupabaseAuth)**: For authenticated users
       - Automatically included when logged in through the web interface
       - Managed by Supabase Auth

    ## Rate Limiting

    All endpoints are rate-limited to prevent abuse:

    - `/api/auth/save-token` & `/api/cron/sync`: Configurable via environment variables
    - `/api/backend/*`: Configurable via `PROXY_RATE_LIMIT_REQUESTS` and `PROXY_RATE_LIMIT_WINDOW`
    - Contact Form (Server Action): Configurable via environment variables

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Base URLs

    - Production: Configured via `NEXT_PUBLIC_APP_URL`
    - Development: `http://localhost:3000`

    ## Error Handling

    All error responses follow a consistent format with appropriate HTTP status codes and descriptive messages.

  contact:
...

...
servers:
...
paths:
...
  /api/profile:
    get:
      tags:
        - Profile
      summary: Fetch user profile
      security:
        - SupabaseAuth: []
      description: |
        Returns the current user's profile data from the local database, soft-synced
        with EzyGo. PII fields (`phone`, `gender`, `birth_date`) are AES-256-GCM
        encrypted at rest; the server decrypts them before responding.

        **Full Bundle:** This endpoint returns an "uber-bundle" containing:
        - Basic profile info
        - Academic context (current_semester, current_year)
        - User settings (bunk calculator, target percentage, disabled courses)
        - Terms compliance status
        - Auth bridge token (decrypted ezygo_token)

        **Origin validation:** In production, requests must include a matching `Origin`
        header (or `Sec-Fetch-Site: same-origin`). Cross-site requests are rejected
        with 400/403 as defence-in-depth against CSRF on this read path.

      operationId: getProfile

      responses:
        "200":
          description: Profile returned successfully
          headers:
            Cache-Control:
              schema:
                type: string
              example: "no-store, max-age=0"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Origin header missing (and Sec-Fetch-Site is not same-origin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Origin not allowed (cross-site request rejected in production)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags:
        - Profile
      summary: Update user profile
      security:
        - SupabaseAuth: []
      description: |
        Updates user-editable profile fields. PII values (`gender`, `birth_date`) are
        AES-256-GCM encrypted before being written to the database.

        **CSRF:** Requires the `x-csrf-token` request header.

        **Validation:** `first_name` is required (min 2 chars). All other fields
        are optional; omitting an optional field does **not** clear it.

      operationId: updateProfile

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - first_name
              properties:
                first_name:
                  type: string
                  minLength: 2
                  example: "Aditya"
                last_name:
                  type: string
                  nullable: true
                  example: "Kumar"
                gender:
                  type: string
                  enum: [male, female, other]
                  nullable: true
                birth_date:
                  type: string
                  format: date
                  nullable: true
                  example: "2003-08-15"

      responses:
        "200":
          description: Profile updated; returns saved plaintext values
          content:
            application/json:
              schema:
                type: object
                properties:
                  first_name:
                    type: string
                  last_name:
                    type: string
                    nullable: true
                  gender:
                    type: string
                    nullable: true
                  birth_date:
                    type: string
                    nullable: true
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Invalid CSRF token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Validation failed"
                  details:
                    type: object
        "500":
          description: Database update failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    UserProfile:
      type: object
      description: Decrypted user profile data (PII fields returned as plaintext)
      properties:
        id:
          oneOf:
            - type: string
            - type: integer
          description: EzyGo user ID
        username:
          type: string
          example: "student123"
        email:
          type: string
          format: email
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
          description: Decrypted phone number (AES-256-GCM at rest)
        gender:
          type: string
          nullable: true
          description: Decrypted gender (AES-256-GCM at rest)
        birth_date:
          type: string
          nullable: true
          description: Decrypted birth date in YYYY-MM-DD format (AES-256-GCM at rest)
        avatar_url:
          type: string
          nullable: true
        terms_version:
          type: string
          nullable: true
        terms_accepted_at:
          type: string
          format: date-time
          nullable: true
        current_semester:
          type: string
          nullable: true
        current_year:
          type: string
          nullable: true
        ezygo_token:
          type: string
          nullable: true
        settings:
          type: object
          properties:
            bunk_calculator_enabled:
              type: boolean
            target_percentage:
              type: integer
            disabled_courses:
              type: object
              additionalProperties:
                type: object
                additionalProperties:
                  type: string
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message describing what went wrong
        message:
          type: string
          description: Detailed error message
      example:
        error: "Internal Server Error"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication using CRON_SECRET environment variable.
        Used for automated cron jobs.

        Example: `Authorization: Bearer <CRON_SECRET>`

    SupabaseAuth:
      type: apiKey
      in: cookie
      name: sb-*
      description: |
        Session cookie authentication managed by Supabase Auth.
        Automatically included when users are logged in through the web interface.
        Cookie names typically start with 'sb-' prefix.
...